Merge file holder types
Before making everything parcelable for item restoring, I merged all the file holder classes to simplify most of the calls (which worked). The only problem being left is to figure out what keeps calling the null methods when the file holder is actually a representative for the titles.

Representative classes are, even though they are working, the most ugliest implementation or workaround that should addressed with something more meaningful and understandable. So-called buttons on the list UI is also ugly and needs a deeper examination so that next implementations don't have to follow something that ugly.