Fix #110 & complete implementing new selection API
The new selection API is flexible enough that it is possible to make most of the calls manually from appropriate locations like Fragment.onOptionsItemSelected() and similar. Also now, the content-sharing callbacks save the select content to the database without needing to call ShareActivity, which means we don't need to parse the data to send it through intents and we can put special data that of split apk which may have more than one APK files.