Bound BackgroundService
It used to be that when we wanted to communicate with the BackgroundService (CommmunicationService previously), we needed to send and wait for receiving intents. The main reason was the service being on another process. This process was ensuring that when the UI thread failed it could keep going. However, the application was not good at showing what was going wrong, when an error happened on the service. With this change, the service is always bound after the app first starts and in background the user leaves all the activities and there are tasks still going on. There are still things to do like ensuring the previous ways of showing the state of tasks. Because this is the background service and handles everything going on the app like file transfer or deletion, it will also responsible of handling all them with occasional interference of the UI which is something already possible with the previous versions. Adding that, there is the new Identity class which will ensure that when we are fetching a task, it is what we have started, which will be achieved not by just the comparison of the hashed intent, but by other means like adding more logical statements to it.