Making BackgroundService the spotlight: Migrate all other small tasks
The background service on TrebleShot is an attempt to make background tasks more sophisticated and easy-to-use. The current struggle is to notify the owner that is detached and might not return. The idea is that with the attachable background tasks, one can attach or detach any listener depending on its availability and the background service is always prepared for this. Now, in the context of activities, sometimes it is expected of user to return to the activity so that the task die in peace meaning returning results or showing dialog in an activity context. However, if user doesn't return, it means the task is in stuck state and has no point of doing what it is doing because the owner is no longer interested in the results.

I am trying possible workarounds to get over this. Of course, it should still be sophisticated and should not require any extra work.

The possible solution would be to separate user notifications to two categories which are notifications and dialogs. With notifications, the user is notified when he or she doesn't have a foreground activity of ours, so that he or she can return to the app. The dialogs are shown when the context has an foreground activity that we can attach dialogs to. However, I don't want to rely on dialogs too much because there are pop-up and may make the user angry as they try to steal the active focus. I am not sure what to use other than dialogs though.