Test implementation of message delivery with background service
I haven't come up with a solution that can both be right and easy (error-free) to inform the user. The current way is simply to assume, we can always show dialogs. Yes, probably we can if we use them on certain conditions like the task stopping when user leaves the host activity, but this is against what I am trying to do with the background service. I want it to handle the task if user does not kill the activity intentionally. This is a way of saving the tasks from the Low Memory Killer program of Android OS. Also, this commit includes some changes to liberate standalone database objects like NetworkDevice from its descendants. If I can achieve this with all the objects, I can make them Parcelable which can be put and get from Bundles.