Let the exceptions lead
Exceptions is an option in the workflow. But should they be? It is usually that when a part of task fails, it is not recoverable, but we try to slowly get out of the task which is wrong. First there may be more mess left behind than it could have ever been, second my practises already allows each undos. For instance, we add an object to the database and user interrupts the task right after. We can add a closer after the instertion and it can remove the data added. Or better is, since I am using transactions which if not followed by transactionSuccessful call, are not valid has no impact on the database. Also I am slowly making classes final and their Parcelable implementations.