Move to more specific exceptions for common errors
A task exiting with null value or returning false may be less informative than it can be. Similar to what a previous commit did, exiting as soon as a fatal error occurs may help to avoid adding more checks below of the scope that ensures that the task did not fail. With specific errors thrown, I can show the right error message and take the action.