Don't use exception when reaching to an item on an adapter
With an old change, I made each access to the list items require try-catch ensure that they are not accessed when the list is updating. However, later, I figured, the problem was not the list being updated, it was holder returning NO_POSITION, and because the view was updating. Also, the item list instance that is being access was not final and was synchronized using the method that makes it accessible, which is not how it is done. However, since the list object belonged to the adapter and getItem is defined there, I was able to lock the object temporarily and fix the issue properly. Also, now, some of the layout click methods were given final keyword so that they can be public and cannot be changed, because these methods ensures that whenever the ViewHolder contains an invalid position, the list is not accessed.