Use CoolSocket 1.0.4
With new CoolSocket version, the support for older version is no longer a possibility since new CoolSocket version handles the connections in a different way (presumably faster). This version fixes the funny behavior when reading in a row or writing, which means I can get rid of STUB messages (a workaround to fix this issue).