Fix crash when the app is not in the foreground and starts
Due to the App class requesting the immediate removal of the service, Android System thinks that the foreground service did not call startForeground when it was created. Even if we call that method, this keeps happening if we use stopService method, which, I think, is a bug. To prevent that, I implemented the old method which gracefully closes the service by waking it up again and make it stop itself with stopSelf() method.