web: Reimplement with AndServer library
This change should fix lots of problems that
NanoHTTPD library caused. Also, AndServer has a
better documentation and easier to use mindful
APIs.

Along with that, I now use Bootstrap and its
dark variant in the templates. It increases the
source size, yes, but I don't plan to update it,
so that shouldn't be a huge problem.

Closes #176
Fixes #152
Fixes #149
Closes #124
Fixes #105