extras: Use local data for changelog & ui: Misc.
Removed retrofit2 dependency since it doesn't
support below API 21. With that, I also removed
 built-in updater and contributors section. The
changelog data is now stored as a raw resource,
so it will still work. I also improved how
changelog and crash logs are presented to the
user. Now, they are fragments and aren't using
the base Activity class. Closes #173.

I also improved the about section and it is also
a fragment now (along with 3rd party licenses
section).

Another UI changes are in the OptionFragment of
client picker and now it has a cleaner look to
it.

Finally, I tested on devices running Android 4.4.4
and below. They are functioning normally, so I
don't think we don't have a big obstacle regarding
the stable release.