saf: Implement switching 'save' directory
This adds the 'Storage path' option in the
settings back (although with different
implementation). I tried to make it look as
simple as possible.

I still think I should add the 'internal storage'
(default external storage) as an option, but that
is for another commit.

This commit also includes some cleanups that
I should have done in another commit.