ui: Reimplement the 'Receive' section
This change turns the 'Receive' section into
something useful. It now uses uprotocol's
REQUEST_GUIDANCE feature that sends signal to a
given sender, which in turn allows that sender to
choose the requesting receiver. That is of course
when the sender is actually a sender. If not, the
receiver is shown a message that "sender" was not
sending anything. Also, this all happens on a
single connection even where there are more one
opearation (e.g., it starts with a REQUEST_TRANSFER
and then receives it).

I also made some tweaks to CoolSocket. It can now
leave an ActiveConnection instance tangling after
the child thread that created it exits. This
allows to the same ActiveConnection instance on
different threads without blocking the owner
thread.

Finally, I am now using lifecycle observers to
check whether there is a subscriber that can
handle a file transfer or guidance request.
This is useful because I don't have to worry
if those requests will be visible to the user
when they arrive. For instance, the 'Receive'
section will navigate the user to the transfer
detail when a transfer request arrives.